autoRemoteWwsieForUnity/requirements.txt
waapi_client==0.7.1
